backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic-settings==2.2.1
pydantic[email]==2.7.1
aiosmtplib==3.0.1
email-validator==2.1.1
langchain-groq==0.1.6
python-dotenv==1.0.1
httpx==0.27.0